Break up class-ajax.php
